Refactor admin.py and forms.py in employee and organization apps
